requirements.txt
# ==============================================================================
# Project Sanjay Mk2 - Python Requirements
# ==============================================================================
# Python 3.11 is recommended for Isaac Sim compatibility.
# This file intentionally avoids fabricated hard pins; install latest compatible
# versions from PyPI.
# ==============================================================================

# Core scientific stack
numpy
scipy
PyYAML
matplotlib
transforms3d

# Flight control and MAVLink (optional for Isaac Sim)
# Install only when using PX4 SITL or real hardware; use backend="isaac_sim" without these
mavsdk
pymavlink
grpcio
protobuf

# Simulation and rendering
mujoco
gymnasium
glfw
PyOpenGL

# AI / ML
torch
torchvision
torchaudio
ultralytics
ultralytics-thop

# Model export / runtime
onnx
onnxruntime

# Computer vision
opencv-python
pillow

# Services / APIs
requests
polars
psutil
websockets
aiohttp

# Testing
pytest
pytest-asyncio

# Development tools
black
isort
flake8
mypy
